+++ korma-triangulate-pathfind/src/commonMain/kotlin/com/soywiz/korma/triangle/ShapeArea.kt
@@ -4,9 +4,9 @@ import com.soywiz.korma.geom.shape.*
 import com.soywiz.korma.geom.triangle.*
 import com.soywiz.korma.triangle.triangulate.*
 
-val Shape2d.area: Double
+val Shape2d.area: Float
     get() = when (this) {
-        is Shape2d.Complex -> this.items.sumByDouble { it.area }
+        is Shape2d.Complex -> this.items.sumByDouble { it.area.toDouble() }.toFloat()
         is Shape2d.WithArea -> this.area
-        else -> this.triangulate().sumByDouble { it.area }
+        else -> this.triangulate().sumByDouble { it.area.toDouble() }.toFloat()
     }
