+++ src/commonMain/kotlin/com/soywiz/korma/geom/bezier/SegmentEmitter.kt
@@ -5,12 +5,12 @@ import com.soywiz.korma.geom.*
 object SegmentEmitter {
     inline fun emit(
         segments: Int,
-        crossinline curveGen: (p: Point, t: Double) -> Point,
+        crossinline curveGen: (p: Point, t: Float) -> Point,
         crossinline gen: (p0: Point, p1: Point) -> Unit,
         p1: Point = Point(),
         p2: Point = Point()
     ) {
-        val dt = 1.0 / segments
+        val dt = 1f / segments
         for (n in 0 until segments) {
             p1.copyFrom(p2)
             p2.copyFrom(curveGen(p2, dt * n))
